refactor(core): Fix import path
